Introduce top level swift package + other improvements
- Rename target "WidgetsExtension" -> "Widgets"
- Fix TEST_HOST (when building for macOS platform)
- Exclude "GreenfieldWidgets" target when building for macOS platform
- Improve "clear-derived-data" script
- Rename "ProjectKit" -> "SharedKit"
- Fix TEST_HOST
- Remove "Frameworks" group
- Rename "Config/Env/Common.xcconfig" to "Shared"
- Add common resources + minor improvements
- Add initial "Common" module
- Add initial "ProjectKit" swift package

+++ Scripts/clear-derived-data.sh
@@ -1,14 +1,22 @@
 #!/bin/sh
 
+XCODE_PROJECT=$(find . -maxdepth 1 -name "*.xcodeproj" | head -1)
+
 DD_PATH=$(
     xcodebuild \
-    -project Greenfield.xcodeproj \
+    -project $XCODE_PROJECT \
     -showBuildSettings \
     | grep -m 1 "CONFIGURATION_BUILD_DIR" \
     | grep -oEi "\/.*" \
-    | rev | cut -d'/' -f4- | rev
+    | rev | cut -d '/' -f4- | rev \
 )
 
-echo "removing directory at path: $DD_PATH"
+echo "ℹ️  derived data path: $DD_PATH"
 
-rm -rf "$DD_PATH"
+read -p "⚠️  clear derived data? [y/n] " -n 1 -r
+echo
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+    rm -rf "$DD_PATH"
+    echo "✅ done!"
+fi
